Update controller logic and integrate database operations for event management
