+++ src/main/java/com/example/pupsis_main_dashboard/controllers/SchoolCalendarController.java
@@ -1,17 +1,26 @@
 package com.example.pupsis_main_dashboard.controllers;
 
-import com.example.pupsis_main_dashboard.utility.SchoolEventLoaderDatabase;
+//import com.example.pupsis_main_dashboard.utility.DBConnection;
+import com.example.pupsis_main_dashboard.databaseOperations.DBConnection;
 import javafx.fxml.FXML;
+import javafx.geometry.Orientation;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.layout.*;
 import javafx.concurrent.Task;
+import javafx.stage.Modality;
+import javafx.stage.StageStyle;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.time.*;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 
-public class SchoolCalendarController extends SchoolEventLoaderDatabase {
+public class SchoolCalendarController {
     @FXML private GridPane calendarGrid;
     @FXML private VBox vBox;
     @FXML private GridPane monthPicker;
@@ -25,6 +34,11 @@ public class SchoolCalendarController extends SchoolEventLoaderDatabase {
     private LocalDate activeMonthDate = LocalDate.now();
     private int currentYear = activeMonthDate.getYear();
     private Month currentMonth = activeMonthDate.getMonth();
+    
+    // Properties from SchoolEventLoaderDatabase
+    protected final Map<String, List<String>> eventsMap = new HashMap<>();
+    protected double xOffset = 0;
+    protected double yOffset = 0;
 
     private final String stylesheetPath = Objects.requireNonNull(getClass().getResource("/com/example/pupsis_main_dashboard/css/SchoolCalendar.css")).toExternalForm();
 
@@ -45,9 +59,134 @@ public class SchoolCalendarController extends SchoolEventLoaderDatabase {
         rightButton.setOnMouseClicked(_ -> handleNextButton());
     }
 
+    // Methods integrated from SchoolEventLoaderDatabase
+    public void loadSchoolEvents() {
+        eventsMap.clear(); // Clear old data before loading new
 
-     // Load school events in the background. This method uses a Task to load school
-     // events asynchronously, and a new thread is started to run the task.
+        String query = """
+                    SELECT ed.event_date, e.event_description, e.event_type
+                    FROM school_dates ed
+                    JOIN school_events e ON ed.event_id = e.event_id
+                """;
+
+        try (Connection conn = DBConnection.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(query);
+             ResultSet rs = stmt.executeQuery()) {
+
+            while (rs.next()) {
+                String date = rs.getString("event_date");
+                String eventName = rs.getString("event_description");
+                String eventType = rs.getString("event_type");
+
+                String formatted = String.format("%s,%s,%s", date, eventType, eventName);
+
+                eventsMap.computeIfAbsent(date, _ -> new ArrayList<>()).add(formatted);
+            }
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void showEventDay(String selectedDate, javafx.scene.Parent root) {
+        // Retrieve all events for the selected date
+        List<String> eventsForDate = eventsMap.getOrDefault(selectedDate, new ArrayList<>());
+
+        StringBuilder description = new StringBuilder();
+
+        if (eventsForDate.isEmpty()) {
+            description.append(selectedDate).append(",No Event,No events for this date.");
+        } else {
+            // Add each event's details to the description
+            for (String event : eventsForDate) {
+                description.append(event).append("\n\n"); // Adding a newline between events
+            }
+        }
+
+        // Call the method to show the event dialog with all the events for this date
+        showEventDialog(description.toString(), root);
+    }
+
+    public void showEventDialog(String description, javafx.scene.Parent root) {
+        String[] eventDetails = description.split("\n\n");  // Split the description into individual events
+
+        Dialog<Void> dialog = new Dialog<>();
+        dialog.initModality(Modality.APPLICATION_MODAL);
+        dialog.getDialogPane().setPrefSize(270, 270);
+        dialog.initStyle(StageStyle.TRANSPARENT);
+        dialog.getDialogPane().getStylesheets().add(Objects.requireNonNull(getClass().getResource("/com/example/pupsis_main_dashboard/css/SchoolCalendar.css")).toExternalForm());
+
+        if (root.getStyleClass().contains("dark-theme")) {
+            // Apply the dark theme to the root scene of the dialog
+            dialog.getDialogPane().getScene().getRoot().getStyleClass().addAll("dark-custom-dialog", "dark-theme");
+        }
+
+        dialog.getDialogPane().getStyleClass().add("custom-dialog");
+        VBox content = new VBox(10);
+
+        // The first event's date will be used as the header
+        String[] firstEvent = eventDetails[0].split(",");
+        LocalDate date = LocalDate.parse(firstEvent[0]);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MM/dd/yyyy");
+        String formattedDate = date.format(formatter);
+
+        // The first event's type will be used as the header
+        String[] firstEventType = eventDetails[0].split(",");
+        String formattedEventType = firstEventType[1];
+
+
+        // Date Label
+        Label dateHeader = new Label(formattedDate);
+        dateHeader.getStyleClass().add("custom-dialog-date");
+
+        Label eventTypeHeader = new Label(formattedEventType);
+        eventTypeHeader.getStyleClass().add("custom-dialog-header");
+
+        // Create a VBox to hold all events for the selected date
+        VBox eventList = new VBox(5);  // Space between event items
+        for (String event : eventDetails) {
+            String[] elements = event.split(",");
+            String eventDescription = elements[2];
+
+            // Event Description Label
+            Label eventDescLabel = new Label(eventDescription);
+            eventDescLabel.getStyleClass().add("custom-dialog-description");
+            eventDescLabel.setWrapText(true);
+
+            // Add an event type and description to the event list
+            eventList.getChildren().add(eventDescLabel);
+        }
+
+        // Separator line
+        Separator separator = new Separator();
+        separator.setOrientation(Orientation.HORIZONTAL);
+
+        // Add all content into the VBox
+        content.getStyleClass().add("custom-dialog-content");
+        content.getChildren().addAll(dateHeader, eventTypeHeader, separator, eventList);
+
+        dialog.getDialogPane().setContent(content);
+        dialog.getDialogPane().getButtonTypes().addAll(ButtonType.OK);
+        dialog.getDialogPane().getScene().setFill(javafx.scene.paint.Color.TRANSPARENT);
+
+        Button okButton = (Button) dialog.getDialogPane().lookupButton(ButtonType.OK);
+        okButton.getStyleClass().add("custom-dialog-button");
+
+        dialog.getDialogPane().setOnMousePressed(event -> {
+            xOffset = dialog.getDialogPane().getScene().getWindow().getX() - event.getScreenX();
+            yOffset = dialog.getDialogPane().getScene().getWindow().getY() - event.getScreenY();
+        });
+
+        dialog.getDialogPane().setOnMouseDragged(event -> {
+            dialog.getDialogPane().getScene().getWindow().setX(event.getScreenX() + xOffset);
+            dialog.getDialogPane().getScene().getWindow().setY(event.getScreenY() + yOffset);
+        });
+
+        dialog.showAndWait();
+    }
+
+    // Load school events in the background. This method uses a Task to load school
+    // events asynchronously, and a new thread is started to run the task.
     private void loadEventsInBackground() {
         Task<Void> preloadEventsTask = new Task<>() {
             @Override
