+++ src/main/resources/com/example/pupsis_main_dashboard/fxml/newGradingModule.fxml
@@ -16,7 +16,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="720.0" prefWidth="810.0" styleClass="root" stylesheets="@../css/FacultyInput.css" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.GradingModule.GradingModuleController">
+<AnchorPane prefHeight="720.0" prefWidth="810.0" styleClass="root" stylesheets="@../css/FacultyInput.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.pupsis_main_dashboard.controllers.GradingModuleController">
    <children>
       <BorderPane prefHeight="720.0" prefWidth="810.0">
          <top>
@@ -97,7 +97,7 @@
                         </Pane>
                         <AnchorPane prefHeight="400.0" prefWidth="750.0">
                            <children>
-                              <TableView fx:id="subjectsTable" id="mainGradingTable" prefHeight="460.0" prefWidth="710.0">
+                              <TableView id="mainGradingTable" fx:id="subjectsTable" prefHeight="460.0" prefWidth="710.0">
                                 <columns>
                                   <TableColumn fx:id="yearSecCol" editable="false" maxWidth="141.0" minWidth="141.0" prefWidth="141.0" text="Year &amp; Section" />
                                   <TableColumn fx:id="semCol" editable="false" maxWidth="110.0" minWidth="110.0" prefWidth="110.0" text="Semester" />
